requirements.txt
fastapi
uvicorn[standard]
requests
pandas
openai
python-dotenv
google-generativeai
pillow